modify the unit test data type error in file mlp.py (and also assert_true in file test_data.py) when setting theano flag to floatX=float32

+++ sknn/tests/test_data.py
@@ -57,13 +57,11 @@ class TestNetworkParameters(unittest.TestCase):
         
         weights = numpy.random.uniform(-1.0, +1.0, (16,4))
         biases = numpy.random.uniform(-1.0, +1.0, (4,))
-        weights = weights.astype(numpy.float32)
-        biases = biases.astype(numpy.float32)
         nn.set_parameters([(weights, biases)])
         
         p = nn.get_parameters()
-        assert_true((p[0].weights == weights).all())
-        assert_true((p[0].biases == biases).all())
+        assert_true((p[0].weights == weights.astype(numpy.float32)).all())
+        assert_true((p[0].biases == biases.astype(numpy.float32)).all())
 
     def test_LayerParamsSkipOneWithNone(self):
         nn = MLPR(layers=[L("Sigmoid", units=32), L("Linear", name='abcd')])
@@ -72,13 +70,11 @@ class TestNetworkParameters(unittest.TestCase):
         
         weights = numpy.random.uniform(-1.0, +1.0, (32,4))
         biases = numpy.random.uniform(-1.0, +1.0, (4,))
-        weights = weights.astype(numpy.float32)
-        biases = biases.astype(numpy.float32)
         nn.set_parameters([None, (weights, biases)])
         
         p = nn.get_parameters()
-        assert_true((p[1].weights == weights).all())
-        assert_true((p[1].biases == biases).all())
+        assert_true((p[1].weights == weights.astype(numpy.float32)).all())
+        assert_true((p[1].biases == biases.astype(numpy.float32)).all())
 
     def test_SetLayerParamsDict(self):
         nn = MLPR(layers=[L("Sigmoid", units=32), L("Linear", name='abcd')])
@@ -87,10 +83,8 @@ class TestNetworkParameters(unittest.TestCase):
         
         weights = numpy.random.uniform(-1.0, +1.0, (32,4))
         biases = numpy.random.uniform(-1.0, +1.0, (4,))
-        weights = weights.astype(numpy.float32)
-        biases = biases.astype(numpy.float32)
         nn.set_parameters({'abcd': (weights, biases)})
         
         p = nn.get_parameters()
-        assert_true((p[1].weights == weights).all())
-        assert_true((p[1].biases == biases).all())
+        assert_true((p[1].weights == weights.astype(numpy.float32)).all())
+        assert_true((p[1].biases == biases.astype(numpy.float32)).all())
